Add volatile annotations to some asm statements.
For whatever reason GCC 11 through 13 will produce an incorrect result
for pcurves P-521 without these volatile qualifiers. I do not know if
this is caused by a bug in our asm constraints or a bug in GCC.

Unfortunately as it only manifests when in amalagamtion mode it is
quite difficult to reduce to a small testcase.